Complete the status transition matrix.
Wind is now 1:1 propagated. This is priliminary, but will require more
thinking, and probably adjustements. 
I do not yet know how to fuse ground speed and direction, wind, and air
speed and heading. My goal is to couple gound and air vector with the
wind as loosely as possible, to avoid wild swings in the wind indication
due to gusts if I directly calculate the wind here in the transition
matrix from the difference between ground and air vector.