Start remodel TAS-Wind-ground vector
Goal is to model the TAS and wind interdependency with ground course and
speed.
At the same time there are the two paths of the cartesian speed vecros
with accelerations and the cylindric vectors with absolute speed and
turn rate which I want to model together
in order that true sensor fusion is achieved.