Correct roll angle for hanging wing, and flying turns
Start implementation of correction of the roll (bank) angle by
implementing a 'rubberband' effect of a hanging wing (static), and
flying a coordinated turn (dynamic).
The correction factor is driven by the difference of the current model
and the calculated value of static and dynamic bank, and configuable
time constants for both values, and the time difference of the current
propagation cycle.
Added the configurable time constants as static members to
GliderBarioTransitionMatrix class.
Added an epic rant as Doxy doc in the transion method.
Started implementing the correction.