Add lastPressure to GliderVarioStatus.
lastPressure is the calculated pressure from the latest altMSL from the status.
This is used to calculate the dynamic pressure from TAS, and avoids re-calculating the expensive barometric euqation to obain pressure, and from that the density.
I am accepting the error by adjustments to the pressure from re-calculating the altitude in the meantime. It is neglectable after initialization