Implement fixture class MeasurementUpdaterTest
In the constructor initialize the status to a slight turn starting at
EDHG at 30m/s with a slow clockwise turn of 1deg/s and climb of 1.2m/s
Initialize the error covariant and the status noise matrix.
Run the status propagation for 100 times at 0.1 sec interval, i.e for
10 sec.
This creates the dependencies in the covariance matrix for further tests
of measurement updates.