Implement the Kalman filter as shared library.
Encapsulate the entire Kalman filter, i.e. all of the name space OpenEV in a shared library.
Refine the preprocessor definitions for importing and exporting to specific shared libraries.
Do not re-compile the main sources for the unit test program but utilize the Kalman filter shared library for the unit tests too.