WIP: Longitude and latitude are split into base and offset.
Longitude and latitude have a value range around the world absolute to mm for precise Kalman deltas which cannot be resolved by float.
Therefore the Kalman status and filter uses the offset in m from the next arc second.
 After the filter update an overflow of the offset into the next arc second isnormalized.
The change has started. There are multiple comile errors because I did not refactor the variables but changed the definition only. Each instance os use must manually be updated because the algorithms change.