GliderVarioMeasurementUpdater> Add calc3DMeasureUpdate
Not used at the moment. The measurements are all statistically
independent. Therefore there is no benefit of using this function over
calcSingleMeasureUpdate three times in a row.
However with calcSingleMeasureUpdate I only need a scalar division
whereas for calc3DMeasureUpdate I need to calculate the inverse of a 3x3
matrix which is a lot more effort.