RotationMatrix: Wrong sign of angles
The angles in the rotation matrix must be negative compared to the
references Wikipedia page.
Reason is that the coefficients in the reference are calculated for a
coordinate system where Z points up, and Y points left.
My coordinate system is flipped over around the X axis, thus Z points
down, and Y points left. Yaw and pitch are therefore also reverse.

Therefore the sine values must be negated.