Program chrashed during startup under Windows/Cygwin.
Automatic registration of serial UDP and TCP port drivers lead to a crash
upon startup when running under Cygwin under Windows 10.
Reason seems that constructors of static objects are called in a different order
under Cygwin than under Linux.
Under Cygwin the static objects of loaded DLLs seem to be called first, then the constructors of static objects of the main program.
This lead to adding of port driver register records into a std::map in the main program.
This std::map was obviously not yet constructed.
Solution is to call the register calls from GliderVarioMainPriv::startup.
This is possible because the port drivers are static in the util DLL.
On the other side sensor driver DLLs are loaded at runtime long after all
objects in the main program are constructed.