MPL3115A2: Implement complete driver thread function.
Setup the device for polling mode with continuous single-shot
acquisition. Set 16-times oversampling.
Polling cycle is 100ms.
Pressure and temparature are acquired from the sensor, and converted to
hPa and DegC, and printed to the logger.
In logger debug mode only temperature are printed. More internal details
in trace mode.