GliderVarioMeasurementUpdater::staticPressureUpd(): Update altMSL and QFF separately to ensure stability
Call calcSingleMeasureUpdate() once with measRowT filled only with the
altMSL derivate, and once with measRowT filled with the QFF derivative.
Filling both derivative in one measRowT vector lead to an instability
which lead to a constant increase of the altMSL and QFF variance in the
vario status over time. This instability seems to be systemic, and
occurred for any ratio of measurement variance to system noise
increment.
Separating both ensured covariance stability. Together with a GPS
erceiver QFF was still adjusting to altitude fluctuations of the GPS
receiver slowly as expected.