MS4515Driver: Use calibration data for sensor bias.
When configured read a previously saved bias value from a calibration
data file.
When the initial value is less than 0.2mBar off the old bias value
accept the initial value as new bias, and store it in the calibration
data file. 0.2 mBar is the dynamic pressure of 20km/h at standard MSL
pressure, at 0C.
If the intial value is further off assume an initilization in flight (or
at least strong wind). Initialize the TAS accordingly.