GliderVarioMeasurementUpdater::dynamicPressureUpd: Consider negative speed
At negative speeds force the calculated pressure to go negative. Otherwise an increase in pressure will
result in a further *decrease* of the calculated speed. Therefore use the absolute value of speed once.
The results are horribly inaccurate at higher speeds, but this will occur only close to 0.
In actual flight this is a non-issue because the speed is always positive except for some hard-core
aerobatics, but that is way beyond the design envelope of this instrument.