TE-MEAS-AbsPressureDriver: Complete implementation for MS5803.
The class TE-MEAS-AbsPressureDriver currently implements the conversion
function only for the MS5803.
All the others have slightly different conversion methods albeit the
overall principle is the same. The devil is in the details.
To do: Subclass TE-MEAS-AbsPressureDriver and implement the actual
temperature and pressure conversion in virtual methods specific for each
sensor type.