TE_MEAS_AbsPressure: Implement the MS5803 driver as first driver
Segregate the CRC verification, and pressure conversion as pure virtual
methods in the base class TE_MEAS_AbsPressureDriver.
Imlepement temperature conversion here without second order
compensation, as this method seems to be universal.

There are two CRC calculation methods: One for the 8-pin sensors, and
one for the 4-pin sensors. Also between these two the PROM block has
different sizes (8 or 7 bytes respective).
Implement 2 intermediate classes EightPinDriver and SevenPinDriver.

The pressure conversion in any case and, when required, the temperature
conversion is delegated to a final class.
The first such class is MS5803Driver.
For the other sensor types more final classes are to be implmented.

The static function getNewTE_MEAS_AbsPressureInstance() returns a new
object of a final class based on the driver type.