Add demo programs for IMU tesing and calibration
Add demo programs in /demo_calibration.

One is for the BMX160 sendor board via UDP.
Physically the sensor board connects via PPP over serial.

The other is for the MPU-9150. This IMU is obsolete.
But many stock OpenVario sensor board have then soldered on.
The MPU-9150 is directly connected with I2C.

Both demo programs also emit the measurement data.
There is a helper program "readfile" accompanying each demo program.
It is used to cumulate a long list of measurements collected from the demo
program and calculate the average and standard deviation of each measurement.
Averaged measurements taken from different orientations of the sensor
are then input into the calculation sheets in /doc/Calc-Calibration*.ods
to set the calibration data for the production openEVario.