DriverBase: Add updateCycle.
Add updateCycle for any driver instance.
For polling drivers it is the actual polling cycle time.
For listening drivers where the connected system determines the cycle
(like GPS receivers) the cycle time is used to determine the variance
increment per cycle.

Add code to try to load the update cycle from the configuration.