Throughout: Stop using NAN and isnan().
isnan is always returning false when compiling with fast-math, i.e. not being
fully IEEE float compliant. This is preferrable for ARM NEON vector units
which are not IEEE compliant, and assume any bit combination as a finite
value.

Also improve the design of reading configurations for drivers minimizing the
use of trigger values but catching exceptions yourself.