src/kalman/ src/main/: Replace NAN and isnan() with finite constant UnInitVal
isnan() always returns false when I compile with -ffast-math, or
-f unsafe-math-optimizations, or -Ofast.
The compiler abandons full IEEE float compliance, and treats any bit pattern
as a finite value.
These optimizations are particularly relevant for the ARM NEON vector unit which
is not IEEE compliant. Without any of the obove options the compiler would
not generate NEON instructions which are benefitial for the extensive matrix
calculations in this program.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>