GliderVarioMeasurementUpdater: Make helpers public, improve inverse
Make the helper functions public. They are static, no internal status
can be messed up.
They become subject to unit testing.

GliderVarioMeasurementUpdater::calcInverse3D: Determinants may be != 0
due to rounding errors when it actually should be 0.
rate an absolute value of < 0.000005 as 0, and the inverse can therefore
not be calculated.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>