GliderVarioStatus: Replace stream functions by io manipulators for specialized matrixes.
The classes GliderVarioStatus::StatusVectorType and
GliderVarioStatus::StatusCoVarianceType are just re-names of
instanciated Eigen matrix classes. Other classes with the same
instanciation are printed by the stream functions for these
GliderVarioStatus classes.
Therefore specialize the printing of GliderVarioStatus::StatusVectorType
and GliderVarioStatus::StatusCoVarianceType by IO manipulators
printStatusVector and printCovMatrix which have objects of the matrix
classes as parameters.