drivers:IGCReaderDriver: Refactor and re-write calcInitialHeading
Refactor calcInitialHeading to findFirstValidRecord as void function.
It fills the new class member firstRecInfo of struct FirstRecordInfo.
The info therein is used to initialize the Kalman status,
and to determine which B/record in the logger file to use as starting
point.
The first valid B-record is the one where the vehicle exceeds 10m/s for
the first time. At this time heading and speed are reliable enough that
the filter starts operating a reasonable stable status without yanking
it violently around leading the Kalman status locking into absurd
statues like going backwards, or entering a super-sonic narrow circle
making a full turn at the time time for each new logger entry, e.g. 2
seconds.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>