GliderVarioTransitionMatrix::calcTransitionMatrixAndStatus: Simplify and correct
Simplify and correct particularly the use of acceleration for vertical and
horizontal speed along the X axis only, but not for positions.
Set cross acceleration to 0. There is no speed tracked along the Y axis, and
no measurement can reign in the ambiguity between an acceleration, and
gravitie's pull when banked. Moreover accelerations accross to flight direction
are rare, and transient only neither affecting speeds nor positions
significantly.