MS4515Driver, IMUBase: Adjust variances of dynamic pressure and accelrometers
Tune the Kalman filter: Let the speed readings correct the pitch angle faster.
The X axis accelerometer reading is defined by real acceleration, resulting
in speed changes, and by the pitch angle and gravity pull even when not
accelerating.
Let the dynamic pressure (i.e. speed reading) correct the acceleration, and
be a corrective for the pitch angle indirectly.