GliderVarioTransitionMatrix: Damp acceleration values by 10% per cycle
Acceleration is transient. Buit the model locks in on an invalid attitude,
and non-existing acceleration mixing up gravity with actual acceleration.

The factors are still far from ideal. Acceleration has way to much effect,
and is not reigned in properly by air speed, and alitmeter.
Accelerometer measurements are still taken too much into account for
acceleration, and not eough for attitude adjustment.