GliderVarioDriverList: Use DriverBase::readCommonConfiguration
GliderVarioDriverList::loadDriverInstances:
- call DriverBase::readCommonConfiguration
- remove reading cycleTime parameter (moved to readCommonConfiguration)

GliderVarioDriverList::loadDriverInstances:
Proper error output when a driver instance failed to load or configure.

GliderVarioDriverList::initDrivers:
Obtaining the pointer to the I/O port moved to driverInit, and can now
throw an exception. When that happens either delete the driver and
continue initializing the remaining devices, or throw the exception
forward and thus terminate the program, based on the configuration
"terminateOnDriverLoadError".