DifferentialPressureSensorBase: Implement calibration data file
Implement reading writing pressure bias as calibration data from and
to a calibration file.

As with all calibration data implmentations it supports a permanent
calibration data file, as well as a dynamically updated one.
Usually the dynamic calibration data file takes precedence.

In ontrast to IMUs where calibration data, i.e. Biases are continuously
refined by the Kalman fiter, and the file is also continuously updated,
in this case the zero-offset is determined after startup, assuming that
the plan it at rest and in calm conditions on the ground.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>

Update to latest Eclipse version 03-2024

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>