AMS5915Driver, MS4515Driver: Use logger of base class
logger is defined as class member in DifferentialPressureSensorBase.
Set the member in the driver classes with their driver names.
Thus all output from class DifferentialPressureSensorBase is done under
the name of AMS5915Driver or MS4515Driver.

In addition in MS4515Driver member i2cAddress was re-defined. It is now
initialized in the constructor body.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>