BusDeviceSensorBase: Add generic processingMainLoop() and helpers
driverThreadFunction() and processingMainLoop() are implemented here,
and will deal with port device, and bus device specific errors.

processOneMeasurementCycle() is now the pure virtual function that must
be implemented per sensor class.

setupSensor() is used to initialize and set the proper operation mode of
the sensor. Per default the method does nothing.

Signed-off-by: Kai Horstmann <horstmannkai@hotmail.com>